[declarative] Re-instating Event type as system-provided archetype (as per original DevDeclarative spec) and not as a custom type (custom events made off this archetype will be the custom types). Moving Event descriptor from custom.xml to application-side (off which the XML version in ui.xml is regenerated). Corrects issue with inability to modify events, reported at UI tests at r10027.
Some improvements in debug logging of applying of security, features to domain types.

Making the various states attribute of schedule viewlets to be a property such that it can be evaluated later than when class body is executed (to avoid any timing issues with workflow loading, on which it depends).