Some adjustments to mapper configuration for domain models DocVersion and AttachmentVersion, and the types they inherit from, namely Version and Change, disabling all polymorphic-related declarations; seems DocVersion and AttachmentVersion records are now always instantiated with the correct domain type.
Adding AttachmentVersions.doc_versions mapper property, as the (many-to-many) inverse of DocVersion.attachment_versions.

Renaming DocVersion "files" container attribute to "filevers", for better clarity (a DocVersion instance may only relate to AttachmentVersions, not to Attachments themselves); UI views for such listing need to be reworked.

Adding a simple experimental mechanism (alchemist.container, string_key) to allow for a more useful string key for instances, that would be independent of db PK identity but still uniquely identifies the (at least within the scope of the container). Note this key is part of public URLs, so part of public API. 
Using above mechanism in AttachmentVersion, that now defines a string_key() method. 

Correcting an error in getFields, that sometimes was failing for names that do not correspond to actual (interface) fields on the type.

Other src cleanup, and comments.