[draft refactor] Initial refactor of workflow states "draft" and "working_draft" down to single state "draft".
Adding a bungeni-system role "bungeni.Drafter".

Creating a utility "assign_ownership" that is to be executed on_cerate of all docs, to assign the role "Drafter" to the actual user who is drafting the doc, and the role "Owner" to the user that is set as the conceptual owner of the doc; replacing previously executed utility "assign_role_owner_to_login". Externalising repeated code to another utility "_determine_related_user", and adjusting the (sonn to be unnecessary) condition "user_is_context_owner" to also use it.

Reworking the "assembly_queston" type:
- dropping the "working_draft" state (temporarily marking the state it as obsolete, to not break with current data sets).
- dropping (conditional) transition to "working_draft"; dropping now-unnecessary condition on state "draft".
- dropping transition "working_draft" to "submitted".
- switching draft/redraft transition permission from "Owner" to "Drafter".
- dropping the facet "draft_Clerk".
- renaming facet "draft_Owner" to "draft" and facet "redraft_Owner" to "redraft", and switching permission grants to "Drafter" (from "Owner").
- switching draft/redraft workspace tab settings from "Owner" to "Drafter".