[group owner] Switching domain.Group to only IDrafted instead of IOwned (i.e. has a drafter but not an owner); should correct excessive privileges gained by every member of a chamber on all the docs in the chamber.
Reverting the assigning of Owner (as the group principal itself) roles on group (added in r11201) as that role will be gained by every member of the group and applies to all items "contained" in the group e.g. docs, thus interfering with Owner role assignments on docs.

Making domain.User and domain.Signatory IOwned, and adding "owner" and "drafter" ORM properties.

Adding utils.misc.capit utility to capitalize space separated words in a phrase.