[SCHEMA CHANGE, subscribe] Correcting failures occuring when subscribing a doc, one being incorrect setup of the user_subscriotion instance, and the other being that one of the (polymorphic identity) columns, "doc_principla.activity", was too short for the string value length -- for this updated the length to 128.
Taking the opportunity to normalize the length of all other CHAR VARYING columns used for polymorphic identity of a sub-type to 128. The following columns were changed:

    "change.action", from 16 to 128
    "audit.audit_type", from 30 to 128
    "principal.type", from 30 to 128
    "member.member_type", from 30 to 128
    "doc_principal.activity" from 16 to 128

Adding alembic schema and data migration script, run: 
    cd /opt/bungeni/bungeni_apps/bungeni
    svn up ./data
    ./bin/alembic upgrade head

Logging error message when unsubscribe on a doc fails.

Correcting type in r11342, in the title for state "transferred_senate", in workflows/assembly_motion.xml